flow: change timeout logic for TCP flow
When a flow is TCP, the normal way of seeing it dead his through
packet such as RST or FIN. The timeout is just here to help and
there is no reason it should be enforce. So if a packet arrives
on a TCP flow that should have timeout it is highly likely to be
a packet coming on this flow.

So let's avoid to timeout a TCP flow when we have a packet coming.

On the performance side, the timeout check was really heavy and
it seems we have a big improvement in packet drops when using
this logic.