eve/alert: log pcap filename
There is a need to log pseudo packets because of the streams logging
so the condition was wrong. We could have thought about another logging
condition but in fact there is nothing possible. The flow could be
almost closed and there could still be some extra packets coming later
which means we could have log coming later without knowing at the
point in time.